Add Graph draft

+++ apps/builder/pages/api/typebots/[id].ts
@@ -1,4 +1,3 @@
-import { Typebot } from '.prisma/client'
 import prisma from 'libs/prisma'
 import { NextApiRequest, NextApiResponse } from 'next'
 import { getSession } from 'next-auth/react'
@@ -11,6 +10,12 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
     return res.status(401).json({ message: 'Not authenticated' })
 
   const id = req.query.id.toString()
+  if (req.method === 'GET') {
+    const typebot = await prisma.typebot.findUnique({
+      where: { id },
+    })
+    return res.send({ typebot })
+  }
   if (req.method === 'DELETE') {
     const typebots = await prisma.typebot.delete({
       where: { id },
@@ -18,7 +23,7 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
     return res.send({ typebots })
   }
   if (req.method === 'PATCH') {
-    const data = JSON.parse(req.body) as Partial<Typebot>
+    const data = JSON.parse(req.body)
     const typebots = await prisma.typebot.update({
       where: { id },
       data,
